refactor(maps): extract shared MapTiler module
MAPTILER_KEY, ensureMaptilerSdk(), hideOutdoorTrails() were copy-pasted
across all three map controllers (guess_map, image_map, results_map).
Triple-duplication was already a maintenance smell pre-PR; the recent
API key inlining made every key rotation a 3-file edit.

Move them to app/javascript/lib/maptiler.js and import where needed.
Also adds an escapeText() helper that both image_map and results_map
already needed for Image#title (XSS surface) — replaces the inline
String(s).replace(/</g, "&lt;") in both files with a single named import.

Importmap pin_all_from "app/javascript/lib" mirrors the existing
controllers pin so other shared helpers can drop in without further
config.

No behavior change.
